Improve idempotency
Prior to this change, several tasks were not performed conditionally,
leading to several tasks being marked as "changed" by ansible when no
change was technically needed in order to acheive a desired
configuration state. If merged, this commit will ensure that after
initial deployment, if no changes are made to role dependencies, no
changes will be made or annotated by the ansible-playbook command.