Add GMan to docker-compose
Adds Gman to the deployment by pulling from the GMan repo
and building a docker image. We then add the GMan service
to the docker-compose file.

A fix to the docker-compose file had to be implemented which
changed the location of the file. Originally we were pointed
at a docker-compose file in pidev's home directory which wasn't
being used in the role. Now we directly modify the openfaas/faas
docker-compose file.